requirements.txt
audioread==3.0.0
auraloss==0.2.1
gin-config==0.4.0
librosa==0.8.0
matplotlib==3.6.2
Pillow==9.3.0
pytorch-lightning==1.1.2
PyYAML
soundfile==0.11.0
scikit-learn==1.1.3
torch==1.13.0
torchaudio==0.13.0
torchcrepe==0.0.12
torchlibrosa==0.0.9
tqdm==4.64.1
zipp==3.10.0
cffi==1.15.1
Cython==0.29.32
numpy==1.23.4
pycparser==2.21
pyliblo3==0.13.0
scipy==1.9.3
soundfile==0.11.0
wandb==0.13.5

